+++ recipes/web_conf.rb
@@ -1,7 +1,7 @@
 # encoding: utf-8
 # Cookbook Name:: chef_splunk
 # Recipe:: search
-# Author: Bernardo Macias <bmacias@httpstergeek.com>
+# Author: Bernardo Macias <bmacias84@gmail.com>
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.